Moved api handling into seperate file to allow testing
Added significant comments.
Added test file for populating with tests at a later time.
