+++ smap/routes/api.js
@@ -1,43 +1,35 @@
+//The file routes all API requests
+
 var express = require('express');
 var router = express.Router();
 var createError = require('http-errors');
+var handler = require('./imports/apiHandler');
 
-var dummyData = {crime_rate: {NV: 1, OR: 2}, salary: {NV: 5, OR: 7}, gdp: {NV: 8, OR: 25}};
-
-/* TODO: Add API documentation. */
-
-/* everything in this file is server responses to requests from the client */
+//The root of the api url offers nothing
+//Potentially we could offer something like API documentation here
 router.get('/', function(req, res, next) {
   next(createError(404));
 });
 
 // response to request for categories available
 router.get('/cats', function(req, res, next) {
-  res.json(['crime_rate', 'salary', 'gdp']);
+  res.json(handler.getCats());
 });
 
 // response to request for data for specific categories, as specified by URL encoding
 router.get('/data', function(req, res, next){
-  var cats = req.query.cat;
-
-  if (cats === undefined){
+  //Get category list, or error if not provided
+  let cats = handler.parseDataURL(req.query);
+  if (!cats){
     next(createError(400));
     return;
   }
 
-  if (!Array.isArray(cats)){
-    //making it so that even a single item is an array we can deal with
-    cats = [cats];
-  }
-
-  let contents = {};
-
-  for (category of cats){
-    if (dummyData[category] === undefined){
-      next(createError(404));
-      return;
-    }
-    contents[category] = dummyData[category];
+  //Gets data associated with those categories, or throws 404 if any category does not exist
+  let contents = handler.getData(cats);
+  if(!contents){
+    next(createError(404));
+    return;
   }
 
   // return contents, a object (dictionary) containing the key/value pairs of requested categories
@@ -45,4 +37,3 @@ router.get('/data', function(req, res, next){
 });
 
 module.exports = router;
-
