+++ smap/tests/apiHandler.test.js
@@ -1,0 +1,5 @@
+var handler = require('../routes/imports/apiHandler');
+
+test('TODO', () => {
+  expect(handler.getCats()).toHaveProperty("length");
+})
